ui: put space in between input/output on odd terminal width
When the terminal width isn't even, there is a space between the right
border of the output box and the end of the terminal, which causes it to
not be aligned to the right border of the prompt box.

Add a space in between the input and output boxes to make the border of
the output box be aligned to the prompt box.